refactor(restaurant): deliveryFee, avgRating 자료형 변경

+++ src/main/java/com/flab/woowahaneats/domain/restaurant/domain/Restaurant.java
@@ -3,8 +3,6 @@ package com.flab.woowahaneats.domain.restaurant.domain;
 import lombok.Builder;
 import lombok.Getter;
 
-import java.math.BigDecimal;
-
 @Getter
 @Builder
 public class Restaurant {
@@ -14,8 +12,8 @@ public class Restaurant {
     private String description;
     private String address;
     private int minOrderAmt;
-    private BigDecimal deliveryFee;
+    private int deliveryFee;
     private String region;
     private boolean open;
-    private BigDecimal avgRating;
+    private double avgRating;
 }
